doc_assistant_csv/requirements.txt
langchain==0.3.6
langchain-community
langchain-core
langchain-text-splitters
langchain-openai
langchain-experimental==0.3.2
langchainhub
langsmith
langgraph
langchain-pinecone
pypdf
faiss-cpu
requests
black
openai
google-search-results
python-dotenv
pinecone
beautifulsoup4
streamlit
firecrawl-py
altair<5
qrcode
pygame
pandas
tabulate
